remove advanced settings, keep mapping components
