--- projects/back-office/src/app/dashboard/pages/advanced-settings/advanced-settings.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { AdvancedSettingsComponent } from './advanced-settings.component';
-
-describe('AdvancedSettingsComponent', () => {
-  let component: AdvancedSettingsComponent;
-  let fixture: ComponentFixture<AdvancedSettingsComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [AdvancedSettingsComponent],
-    }).compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(AdvancedSettingsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
