+++ projects/back-office/src/app/dashboard/pages/users/users-routing.module.ts
@@ -8,19 +8,6 @@ const routes: Routes = [
     path: '',
     component: UsersComponent,
   },
-  // TODO: rearrange or delete
-  // {
-  //   path: 'advanced-settings',
-  //   loadChildren: () =>
-  //     import('../advanced-settings/advanced-settings.module').then(
-  //       (m) => m.AdvancedSettingsModule
-  //     ),
-  //   data: {
-  //     breadcrumb: {
-  //       key: 'components.users.advancedSettings.title',
-  //     },
-  //   },
-  // },
 ];
 
 /** Users routing module. */
