+++ projects/safe/src/lib/components/mapping/mapping-modal/mapping-modal.component.ts
@@ -1,8 +1,7 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { Mapping } from '../../../models/setting.model';
-import { createFormGroup } from '../mapping-forms';
+import { createFormGroup, Mapping } from '../mapping-forms';
 
 /**
  * Interface for the data injected into the modal.
