+++ projects/safe/src/lib/components/users/users.component.html
@@ -70,15 +70,6 @@
     >
       {{ 'components.users.invite.confirm' | translate }}
     </safe-button>
-    <!-- <safe-button
-      *ngIf="advancedSettings"
-      icon="settings"
-      category="secondary"
-      variant="primary"
-      (click)="onAdvancedSettings()"
-      [title]="'common.advancedSettings' | translate"
-    >
-    </safe-button> -->
   </div>
 </div>
 <div>
