+++ projects/safe/src/public-api.ts
@@ -48,7 +48,6 @@ export * from './lib/models/pullJob.model';
 export * from './lib/models/layout.model';
 export * from './lib/models/aggregation.model';
 export * from './lib/models/reference-data.model';
-export * from './lib/models/setting.model';
 
 // === COMPONENTS ===
 export * from './lib/components/aggregation/edit-aggregation-modal/edit-aggregation-modal.component';
